Memorise when the request is revised
Previously, every time a Response was been revised the `reviseRequest()`
method was called internally. This was because in the `reviseRequest()`
process the applicable revisions are memorise and then used to be applied
to the response. Calling the `reviseRequest()` method internally from the
`reviseReponse()`method ensured the applicable revisions to apply to the
response were known, but at the cost of calling twice the `reviseRequest()`
method in most scenarios: first when revising the request, and second when
revising the response.

To prevent the unnecessary extra call to `reviseRequest()` a boolean
property indicating whether or not the request has already been revised
has been introduced.