Decouple from `MessageRevision`
There were still many places where we were referencing `MessageRevision`
instances in DocBlocks and tests. Here we're replacing these by the new
Request and Response revisions where applicable.

Also a new marker `Revision` interface has been introduced to be able to
reference it in some type hints, return types and DocBlocks.

As part of these changes, the `Brief` does not longer validate that the
revisions are of a specific type and it simply accepts any `Revision`
instances.