Add support for external revision definition resolvers
The main reason for this change is to introduce a point of extension
to allow revision definitions to be resolved out of a container when
used in conjunction with a one.

`SimpleRevisionResolver` is provided as the default revision resolver
to be used when no container is used or required.

As part of this change, revision definitions has been promoted to its
own class `RevisionDefinition`. This means that rather than passing
arround an array of strings and/or closures representing a list of
revision definitions these are now each wrapped into a
`RevisionDefinition` class, allowing us to add parameter and return
types and having a single place where to validate them.