Assume up to date request if no version specified
Instead of assuming the request is the very first version and needs to be sent
through all applicable revisions, this change modifies the behaviour so the
Chief Editor can assume it is the latest version, so he will create the brief
with no revisions.

Integrators should still be able to define their own `VersionResolver`
implementation (which could be just a decorator) to define their custom fallback
version strategy.