Editor track the applicable response revisions only
A list of applicable `ResponseRevision`s is tracked by the Editor when revising
the request. This is to avoid having to find them again when revising the
response.

Previously, the Editor was keeping track of both request and response revisions
even though only response revisions were been used later on. From now on, only
`ResponseRevision`s will be tracked, allowing us to simplify the
`reviseResponse()` method.