Allow string identifiers from `RevisionDefintion`s
With the `PSR11RevisionResolver` now implemented, it makes sense to allow
`RevisionDefinition`s to return any random string as a revision identifier to
resolved out of the container.

String revision identifiers will result on an exception when using the
`SimpleRevisionResolver` and the string is not a FQCN implementing the
`Revision` interface.