Setup PHP CS Fixer
Adds two separate configuration files, `.php_cs.spec.dist` for phpspec files
and `.php_cs.dist` for everything else.

At the moment, the only difference between the two is that the phpspec rules
file does not enforce method visibility, whereas the configuration for the rest
of the source code does enforce it.

PHP CS Fixer is not required as a composer dependency as it is still not
compatible with PHP 8 and therefore causes the tests workflow in GitHub Actions
to fail when installing the composer dependencies in a PHP 8 environment. Also,
as explained in [this comment](https://github.com/FriendsOfPHP/PHP-CS-Fixer/issues/4702#issuecomment-719545818),
PHP CS Fixer should not be required even as a dev dependency, but installed
separately.