✨ Add Version comparison support
Allow devs to, having a version instance, compare it against another version.
For instance, comparing the version retrieved using the `VersionResolver` with
a specific version to apply different logic. This is helpful to apply logic
depending on the target version. Example:
```
if ($currentVersion->isBefore(new Version('foo'))) {
    // Apply custom logic.
}
```

Supported comparisons methods:
* `isBefore`
* `isSameOrBefore`
* `isSame`
* `isAfter`
* `isSameOrAfter`